refactor: Rename `NumericInput` to `SliderNumericInput`
