+++ web/src/components/SliderNumericInput.vue
